requirements.txt
streamlit>=1.32.0
spotipy>=2.23.0
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0 